OD20DEVAT3-26 minor changes Spring security, User, UserDao

+++ src/main/java/com/netcracker/odstc/logviewer/mapper/UserMapper.java
@@ -2,8 +2,6 @@ package com.netcracker.odstc.logviewer.mapper;
 
 import com.netcracker.odstc.logviewer.models.User;
 import org.springframework.jdbc.core.RowMapper;
-
-import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -11,7 +9,7 @@ public class UserMapper implements RowMapper<User> {
 
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-        User user = new User(BigInteger.valueOf(rs.getLong("object_id")));
+        User user = new User(rs.getBigDecimal("object_id").toBigInteger());
         return user;
     }
 }
